Move RegressionModel into sklearn_model.py
