--- darts/models/forecasting/regression_model.py
@@ -1,339 +1,0 @@
-"""
-Regression Model
-----------------
-
-Note: `RegressionModel` is deprecated and will be removed in a future version. Use `SKLearnModel` instead
-(see `SKLearnModel <https://unit8co.github.io/darts/generated_api/darts.models.forecasting.sklearn_model.html#darts.models.forecasting.sklearn_model.SKLearnModel>`_).
-
-A `RegressionModel` forecasts future values of a target series based on
-
-* The target series (past lags only)
-
-* An optional past_covariates series (past lags only)
-
-* An optional future_covariates series (possibly past and future lags)
-
-* Available static covariates
-
-
-The regression models are learned in a supervised way, and they can wrap around any "scikit-learn like" regression model
-acting on tabular data having ``fit()`` and ``predict()`` methods.
-
-Darts also provides :class:`LinearRegressionModel` and :class:`RandomForestModel`, which are regression models
-wrapping around scikit-learn linear regression and random forest regression, respectively.
-
-Behind the scenes this model is tabularizing the time series data to make it work with regression models.
-
-The lags can be specified either using an integer - in which case it represents the _number_ of (past or future) lags
-to take into consideration, or as a list - in which case the lags have to be enumerated (strictly negative values
-denoting past lags and positive values including 0 denoting future lags).
-When static covariates are present, they are appended to the lagged features. When multiple time series are passed,
-if their static covariates do not have the same size, the shorter ones are padded with 0 valued features.
-"""
-
-from typing import Optional, Union
-
-from darts.logging import get_logger, raise_deprecation_warning
-from darts.models.forecasting.sklearn_model import (
-    FUTURE_LAGS_TYPE,
-    LAGS_TYPE,
-    SKLearnModel,
-    SKLearnModelWithCategoricalCovariates,
-)
-
-logger = get_logger(__name__)
-
-
-class RegressionModel(SKLearnModel):
-    def __init__(
-        self,
-        lags: Optional[LAGS_TYPE] = None,
-        lags_past_covariates: Optional[LAGS_TYPE] = None,
-        lags_future_covariates: Optional[FUTURE_LAGS_TYPE] = None,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        add_encoders: Optional[dict] = None,
-        model=None,
-        multi_models: Optional[bool] = True,
-        use_static_covariates: bool = True,
-    ):
-        """Regression Model
-        Can be used to fit any scikit-learn-like regressor class to predict the target time series from lagged values.
-
-        Note: `RegressionModel` is deprecated and will be removed in a future version. Use `SKLearnModel` instead
-        (see `SKLearnModel <https://unit8co.github.io/darts/generated_api/darts.models.forecasting.sklearn_model.html#darts.models.forecasting.sklearn_model.SKLearnModel>`_).
-
-        Parameters
-        ----------
-        lags
-            Lagged target `series` values used to predict the next time step/s.
-            If an integer, must be > 0. Uses the last `n=lags` past lags; e.g. `(-1, -2, ..., -lags)`, where `0`
-            corresponds the first predicted time step of each sample. If `output_chunk_shift > 0`, then
-            lag `-1` translates to `-1 - output_chunk_shift` steps before the first prediction step.
-            If a list of integers, each value must be < 0. Uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `series` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (integer or list of integers). The
-            key 'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        lags_past_covariates
-            Lagged `past_covariates` values used to predict the next time step/s.
-            If an integer, must be > 0. Uses the last `n=lags_past_covariates` past lags; e.g. `(-1, -2, ..., -lags)`,
-            where `0` corresponds to the first predicted time step of each sample. If `output_chunk_shift > 0`, then
-            lag `-1` translates to `-1 - output_chunk_shift` steps before the first prediction step.
-            If a list of integers, each value must be < 0. Uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `past_covariates` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (integer or list of integers). The
-            key 'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        lags_future_covariates
-            Lagged `future_covariates` values used to predict the next time step/s. The lags are always relative to the
-            first step in the output chunk, even when `output_chunk_shift > 0`.
-            If a tuple of `(past, future)`, both values must be > 0. Uses the last `n=past` past lags and `n=future`
-            future lags; e.g. `(-past, -(past - 1), ..., -1, 0, 1, .... future - 1)`, where `0` corresponds the first
-            predicted time step of each sample. If `output_chunk_shift > 0`, the position of negative lags differ from
-            those of `lags` and `lags_past_covariates`. In this case a future lag `-5` would point at the same
-            step as a target lag of `-5 + output_chunk_shift`.
-            If a list of integers, uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `future_covariates` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (tuple or list of integers). The key
-            'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        output_chunk_length
-            Number of time steps predicted at once (per chunk) by the internal model. It is not the same as forecast
-            horizon `n` used in `predict()`, which is the desired number of prediction points generated using a
-            one-shot- or autoregressive forecast. Setting `n <= output_chunk_length` prevents auto-regression. This is
-            useful when the covariates don't extend far enough into the future, or to prohibit the model from using
-            future values of past and / or future covariates for prediction (depending on the model's covariate
-            support).
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future (relative to the
-            input chunk end). This will create a gap between the input (history of target and past covariates) and
-            output. If the model supports `future_covariates`, the `lags_future_covariates` are relative to the first
-            step in the shifted output chunk. Predictions will start `output_chunk_shift` steps after the end of the
-            target `series`. If `output_chunk_shift` is set, the model cannot generate autoregressive predictions
-            (`n > output_chunk_length`).
-        add_encoders
-            A large number of past and future covariates can be automatically generated with `add_encoders`.
-            This can be done by adding multiple pre-defined index encoders and/or custom user-made functions that
-            will be used as index encoders. Additionally, a transformer such as Darts' :class:`Scaler` can be added to
-            transform the generated covariates. This happens all under one hood and only needs to be specified at
-            model creation.
-            Read :meth:`SequentialEncoder <darts.dataprocessing.encoders.SequentialEncoder>` to find out more about
-            ``add_encoders``. Default: ``None``. An example showing some of ``add_encoders`` features:
-
-            .. highlight:: python
-            .. code-block:: python
-
-                def encode_year(idx):
-                    return (idx.year - 1950) / 50
-
-                add_encoders={
-                    'cyclic': {'future': ['month']},
-                    'datetime_attribute': {'future': ['hour', 'dayofweek']},
-                    'position': {'past': ['relative'], 'future': ['relative']},
-                    'custom': {'past': [encode_year]},
-                    'transformer': Scaler(),
-                    'tz': 'CET'
-                }
-            ..
-        model
-            Scikit-learn-like model with ``fit()`` and ``predict()`` methods. Also possible to use model that doesn't
-            support multi-output regression for multivariate timeseries, in which case one regressor
-            will be used per component in the multivariate series.
-            If None, defaults to: ``sklearn.linear_model.LinearRegression(n_jobs=-1)``.
-        multi_models
-            If True, a separate model will be trained for each future lag to predict. If False, a single model
-            is trained to predict all the steps in 'output_chunk_length' (features lags are shifted back by
-            `output_chunk_length - n` for each step `n`). Default: True.
-        use_static_covariates
-            Whether the model should use static covariate information in case the input `series` passed to ``fit()``
-            contain static covariates. If ``True``, and static covariates are available at fitting time, will enforce
-            that all target `series` have the same static covariate dimensionality in ``fit()`` and ``predict()``.
-
-        Examples
-        --------
-        >>> from darts.datasets import WeatherDataset
-        >>> from darts.models import SKLearnModel
-        >>> from sklearn.linear_model import Ridge
-        >>> series = WeatherDataset().load()
-        >>> # predicting atmospheric pressure
-        >>> target = series['p (mbar)'][:100]
-        >>> # optionally, use past observed rainfall (pretending to be unknown beyond index 100)
-        >>> past_cov = series['rain (mm)'][:100]
-        >>> # optionally, use future temperatures (pretending this component is a forecast)
-        >>> future_cov = series['T (degC)'][:106]
-        >>> # wrap around the sklearn Ridge model
-        >>> model = SKLearnModel(
-        >>>     model=Ridge(),
-        >>>     lags=12,
-        >>>     lags_past_covariates=4,
-        >>>     lags_future_covariates=(0,6),
-        >>>     output_chunk_length=6
-        >>> )
-        >>> model.fit(target, past_covariates=past_cov, future_covariates=future_cov)
-        >>> pred = model.predict(6)
-        >>> pred.values()
-        array([[1005.73340676],
-                [1005.71159051],
-                [1005.7322616 ],
-                [1005.76314504],
-                [1005.82204348],
-                [1005.89100967]])
-        """
-        raise_deprecation_warning(
-            "`RegressionModel` is deprecated and will be removed in a future version. "
-            "Use `SKLearnModel` instead.",
-            logger,
-        )
-
-        super().__init__(
-            lags=lags,
-            lags_past_covariates=lags_past_covariates,
-            lags_future_covariates=lags_future_covariates,
-            output_chunk_length=output_chunk_length,
-            output_chunk_shift=output_chunk_shift,
-            add_encoders=add_encoders,
-            model=model,
-            multi_models=multi_models,
-            use_static_covariates=use_static_covariates,
-        )
-
-
-class RegressionModelWithCategoricalCovariates(SKLearnModelWithCategoricalCovariates):
-    def __init__(
-        self,
-        model,
-        lags: Union[int, list] = None,
-        lags_past_covariates: Union[int, list[int]] = None,
-        lags_future_covariates: Union[tuple[int, int], list[int]] = None,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        add_encoders: Optional[dict] = None,
-        multi_models: Optional[bool] = True,
-        use_static_covariates: bool = True,
-        categorical_past_covariates: Optional[Union[str, list[str]]] = None,
-        categorical_future_covariates: Optional[Union[str, list[str]]] = None,
-        categorical_static_covariates: Optional[Union[str, list[str]]] = None,
-    ):
-        """
-        Extension of `RegressionModel` for regression models that support categorical covariates.
-
-        Note: `RegressionModelWithCategoricalCovariates` is deprecated and will be removed in a future version.
-        Use `SKLearnModelWithCategoricalCovariates` instead
-        (see `SKLearnModelWithCategoricalCovariates <https://unit8co.github.io/darts/generated_api/darts.models.forecasting.sklearn_model.html#darts.models.forecasting.sklearn_model.SKLearnModelWithCategoricalCovariates>`_).
-
-        Parameters
-        ----------
-        model
-            Scikit-learn-like model with ``fit()`` and ``predict()`` methods. Also possible to use model that doesn't
-            support multi-output regression for multivariate timeseries, in which case one regressor
-            will be used per component in the multivariate series.
-        lags
-            Lagged target `series` values used to predict the next time step/s.
-            If an integer, must be > 0. Uses the last `n=lags` past lags; e.g. `(-1, -2, ..., -lags)`, where `0`
-            corresponds the first predicted time step of each sample. If `output_chunk_shift > 0`, then
-            lag `-1` translates to `-1 - output_chunk_shift` steps before the first prediction step.
-            If a list of integers, each value must be < 0. Uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `series` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (integer or list of integers). The
-            key 'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        lags_past_covariates
-            Lagged `past_covariates` values used to predict the next time step/s.
-            If an integer, must be > 0. Uses the last `n=lags_past_covariates` past lags; e.g. `(-1, -2, ..., -lags)`,
-            where `0` corresponds to the first predicted time step of each sample. If `output_chunk_shift > 0`, then
-            lag `-1` translates to `-1 - output_chunk_shift` steps before the first prediction step.
-            If a list of integers, each value must be < 0. Uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `past_covariates` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (integer or list of integers). The
-            key 'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        lags_future_covariates
-            Lagged `future_covariates` values used to predict the next time step/s. The lags are always relative to the
-            first step in the output chunk, even when `output_chunk_shift > 0`.
-            If a tuple of `(past, future)`, both values must be > 0. Uses the last `n=past` past lags and `n=future`
-            future lags; e.g. `(-past, -(past - 1), ..., -1, 0, 1, .... future - 1)`, where `0` corresponds the first
-            predicted time step of each sample. If `output_chunk_shift > 0`, the position of negative lags differ from
-            those of `lags` and `lags_past_covariates`. In this case a future lag `-5` would point at the same
-            step as a target lag of `-5 + output_chunk_shift`.
-            If a list of integers, uses only the specified values as lags.
-            If a dictionary, the keys correspond to the `future_covariates` component names (of the first series when
-            using multiple series) and the values correspond to the component lags (tuple or list of integers). The key
-            'default_lags' can be used to provide default lags for un-specified components. Raises and error if some
-            components are missing and the 'default_lags' key is not provided.
-        output_chunk_length
-            Number of time steps predicted at once (per chunk) by the internal model. It is not the same as forecast
-            horizon `n` used in `predict()`, which is the desired number of prediction points generated using a
-            one-shot- or autoregressive forecast. Setting `n <= output_chunk_length` prevents auto-regression. This is
-            useful when the covariates don't extend far enough into the future, or to prohibit the model from using
-            future values of past and / or future covariates for prediction (depending on the model's covariate
-            support).
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future (relative to the
-            input chunk end). This will create a gap between the input (history of target and past covariates) and
-            output. If the model supports `future_covariates`, the `lags_future_covariates` are relative to the first
-            step in the shifted output chunk. Predictions will start `output_chunk_shift` steps after the end of the
-            target `series`. If `output_chunk_shift` is set, the model cannot generate autoregressive predictions
-            (`n > output_chunk_length`).
-        add_encoders
-            A large number of past and future covariates can be automatically generated with `add_encoders`.
-            This can be done by adding multiple pre-defined index encoders and/or custom user-made functions that
-            will be used as index encoders. Additionally, a transformer such as Darts' :class:`Scaler` can be added to
-            transform the generated covariates. This happens all under one hood and only needs to be specified at
-            model creation.
-            Read :meth:`SequentialEncoder <darts.dataprocessing.encoders.SequentialEncoder>` to find out more about
-            ``add_encoders``. Default: ``None``. An example showing some of ``add_encoders`` features:
-
-            .. highlight:: python
-            .. code-block:: python
-
-                def encode_year(idx):
-                    return (idx.year - 1950) / 50
-
-                add_encoders={
-                    'cyclic': {'future': ['month']},
-                    'datetime_attribute': {'future': ['hour', 'dayofweek']},
-                    'position': {'past': ['relative'], 'future': ['relative']},
-                    'custom': {'past': [encode_year]},
-                    'transformer': Scaler(),
-                    'tz': 'CET'
-                }
-            ..
-        multi_models
-            If True, a separate model will be trained for each future lag to predict. If False, a single model is
-            trained to predict at step 'output_chunk_length' in the future. Default: True.
-        use_static_covariates
-            Whether the model should use static covariate information in case the input `series` passed to ``fit()``
-            contain static covariates. If ``True``, and static covariates are available at fitting time, will enforce
-            that all target `series` have the same static covariate dimensionality in ``fit()`` and ``predict()``.
-        categorical_past_covariates
-            Optionally, component name or list of component names specifying the past covariates that should be treated
-            as categorical.
-        categorical_future_covariates
-            Optionally, component name or list of component names specifying the future covariates that should be
-            treated as categorical.
-        categorical_static_covariates
-            Optionally, string or list of strings specifying the static covariates that should be treated as
-            categorical.
-        """
-
-        raise_deprecation_warning(
-            "`RegressionModelWithCategoricalCovariates` is deprecated and will be removed in a future version. "
-            "Use `SKLearnModelWithCategoricalCovariates` instead.",
-            logger,
-        )
-
-        super().__init__(
-            lags=lags,
-            lags_past_covariates=lags_past_covariates,
-            lags_future_covariates=lags_future_covariates,
-            output_chunk_length=output_chunk_length,
-            output_chunk_shift=output_chunk_shift,
-            add_encoders=add_encoders,
-            model=model,
-            multi_models=multi_models,
-            use_static_covariates=use_static_covariates,
-            categorical_past_covariates=categorical_past_covariates,
-            categorical_future_covariates=categorical_future_covariates,
-            categorical_static_covariates=categorical_static_covariates,
-        )
